--- .eslintrc
@@ -1,13 +1,0 @@
-{
-  "parser": "babel-eslint",
-  "env": {
-    "browser": true,
-    "node": true,
-    "jest": true
-  },
-  "plugins": [
-    "react",
-    "react-native"
-  ]
-  #"extends": ["eslint:recommended", "plugin:react/recommended"]
-}
