+++ app.json
@@ -1,7 +1,5 @@
 {
   "expo": {
-    "sdkVersion": "17.0.0",
-    "privacy": "public",
-    "name": "Top UFC Fighters",
-  },
+    "sdkVersion": "17.0.0"
+  }
 }
