+++ package.json
@@ -2,18 +2,9 @@
   "name": "rn-top-ufc-fighters",
   "version": "0.1.0",
   "private": false,
-  "author": {
-    "name": "Joao Marins",
-    "email": "jgcmarins@gmail.com",
-    "url": "https://github.com/jgcmarins"
-  },
   "devDependencies": {
-    "babel-eslint": "^7.2.3",
-    "eslint": "^3.19.0",
-    "eslint-plugin-react": "^7.0.1",
-    "eslint-plugin-react-native": "^2.3.2",
+    "react-native-scripts": "0.0.31",
     "jest-expo": "~1.0.1",
-    "react-native-scripts": "0.0.30",
     "react-test-renderer": "16.0.0-alpha.6"
   },
   "main": "./node_modules/react-native-scripts/build/bin/crna-entry.js",
@@ -29,7 +20,6 @@
   },
   "dependencies": {
     "expo": "^17.0.0",
-    "graphql": "^0.10.1",
     "react": "16.0.0-alpha.6",
     "react-native": "^0.44.0"
   }
