--- src/components/FighterCard.js
@@ -1,30 +1,0 @@
-import React from 'react'
-import { StyleSheet, View } from 'react-native'
-
-export default class FighterCard extends React.Component {
-  render() {
-    return (
-      <View style={styles.container}>
-        {this.props.children}
-      </View>
-    )
-  }
-}
-
-const styles = StyleSheet.create({
-  container: {
-    borderWidth: 1,
-    borderRadius: 2,
-    borderColor: '#ddd',
-    borderBottomWidth: 0,
-    shadowColor: '#000',
-    shadowOffset: { width: 0, height: 2},
-    shadowOpacity: 0.1,
-    shadowRadius: 2,
-    elevation: 1,
-    position: 'relative',
-    marginLeft: 5,
-    marginRight: 5,
-    marginTop: 10,
-  }
-})
